requirements.txt
astroid==2.15.5
colorama==0.4.6
dill==0.3.6
isort==5.12.0
lazy-object-proxy==1.9.0
lxml==4.9.2
mccabe==0.7.0
platformdirs==3.5.1
pylint==2.17.4
python-dateutil==2.8.2
python-docx==0.8.11
six==1.16.0
tomli==2.0.1
tomlkit==0.11.8
typing_extensions==4.6.2
wrapt==1.15.0